requirements.txt
Django==4.2.16
djangorestframework==3.14.0
gunicorn==21.2.0        # para prod (Linux)
whitenoise==6.7.0       # estáticos en prod simple
python-dotenv==1.0.1    # cargar .env en dev
psycopg2-binary==2.9.9  # si migras a Postgres